php.ini-production: disable opcache.huge_code_pages by default
There are only very narrow circumstances under which this option has
been reported to provide 1% performance gain due to reduction of TLB
misses.  In many setups, this option only increases memory usage, and
will actually decrease performance.  To avoid this, let's leave it
disabled by default, and let it be an explicit decision to enable it.

For a discussion, see https://github.com/php/php-src/pull/10301

Closes GH-10336